Updated blender - 17:5.1.2-1

+++ galaxy/blender/PKGBUILD
@@ -9,8 +9,8 @@
 # fine with them.
 
 pkgname=blender
-pkgver=5.1.1
-pkgrel=8
+pkgver=5.1.2
+pkgrel=1
 epoch=17
 pkgdesc="A fully integrated 3D graphics creation suite"
 arch=('x86_64')
@@ -116,6 +116,11 @@ makedepends=(
   'wayland-protocols'
   'vulkan-headers'
 )
+
+makedepends_aarch64=(
+  'sse2neon'
+)
+
 optdepends=('cuda: Cycles renderer CUDA support'
             'intel-compute-runtime: Cycles renderer Intel OneAPI support'
             'intel-level-zero-raytracing-support: Cycles renderer Intel OneAPI Raytracing Support'
@@ -128,7 +133,7 @@ source=("git+https://projects.blender.org/blender/blender.git#tag=v$pkgver"
         blender-hip-update.patch
         blender-fix-oneapi-2026-atomic-address-space.patch::https://raw.githubusercontent.com/intel/llvm/20a7095cba72ace59f7c8a64711ec4b51f01f030/devops/actions/blender/blender-build/patches/Fix-build.patch
         https://developer.download.nvidia.com/redist/optix/v8.0/OptiX-8.0-Include.zip)
-sha512sums=('fd408f920c3762c155fcf47cde18f78800d2565b12dc671836c53f7a43beec782059b42d07856b8b168b37ece02eba6048338ee0076ef7f2724113cd5ee65dc0'
+sha512sums=('1773e1cfe06bf924aea253dcc6c5248ac7f9963c2809598b5e1c18ab2d4a40c2d8630d4a6d55a77e6ef217e11bf4955b389c75a989548a2d1547da9fd5ebd791'
             '77d202e2033a2e5c26adcc5340da6fbd7f859a8b237b37f9be7f08fbbb99173a67462f1b0aa0dce31967cd8c465b6341628567ace1e477b81a1b75c2383357ca'
             '17b15a7e2ea7e89e22a35325104e6a047253242f820815fa499a1010a0caa90b5a7c0a67b4434e24d472d09c767d34a6524e769ca8f1da13728dec89ed9267c0'
             '5502d9df847de12badc702c0444bd4f1f7620460b2235026df2c3133da1e04c148af0f1fc7f345e9a0c009c32f905f66c8d427743445e8864d3a797cdce6a483')
@@ -184,9 +189,6 @@ build() {
     -D OCLOC_INSTALL_DIR=/usr
     -D OPTIX_ROOT_DIR="$srcdir"
     -D PYTHON_VERSION="$(_get_pyver)"
-    -D SYCL_CPP_FLAGS="--verbose"  # for debugging
-    -D SYCL_ROOT_DIR=/opt/intel/oneapi/compiler/latest
-    -D SYCL_OFFLINE_COMPILER_PARALLEL_JOBS=8
     -D USD_ROOT_DIR=/usr
     -D WITH_CYCLES_OSL=ON
     -D WITH_CYCLES_PARALLEL_DEVICE_KERNEL_BUILD=ON
@@ -198,6 +200,22 @@ build() {
     -S "$pkgname"
     -W no-dev
   )
+
+  if [[ $CARCH == "x86_64" ]]; then
+    cmake_options+=(
+      -D SYCL_ROOT_DIR=/opt/intel/oneapi/compiler/latest
+      -D SYCL_CPP_FLAGS="--verbose"  # for debugging
+      -D SYCL_OFFLINE_COMPILER_PARALLEL_JOBS=8
+    )
+  fi
+
+  if [[ $CARCH == "aarch64" ]]; then
+    cmake_options+=(
+      -D WITH_CYCLES_HIP_BINARIES=OFF
+      -D WITH_CYCLES_HIPRT_BINARIES=OFF
+      -D WITH_CYCLES_ONEAPI_BINARIES=OFF
+    )
+  fi
 
   cmake "${cmake_options[@]}"
   cmake --build build
@@ -214,8 +232,10 @@ package() {
   install -Dm644 release/freedesktop/org.blender.Blender.metainfo.xml "${pkgdir}/usr/share/metainfo/org.blender.Blender.metainfo.xml"
 
   # Move OneAPI AOT lib to proper place
-  mkdir "${pkgdir}"/usr/lib/
-  mv "${pkgdir}"/usr/share/blender/lib/libcycles_kernel_oneapi_aot.so "${pkgdir}"/usr/lib/
+  mkdir -p "${pkgdir}"/usr/lib/
+  if [[ -f "${pkgdir}"/usr/share/blender/lib/libcycles_kernel_oneapi_aot.so ]]; then
+    mv "${pkgdir}"/usr/share/blender/lib/libcycles_kernel_oneapi_aot.so "${pkgdir}"/usr/lib/
+  fi
 
   install -vDm 644 doc/license/{BSD-{2,3}-Clause,MIT,Zlib}-license.txt -t "$pkgdir/usr/share/licenses/$pkgname/"
 }
